% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mochi.R
\name{mochi}
\alias{mochi}
\title{Run MoCHI pipeline}
\usage{
mochi(
  workspacePath = NULL,
  inputFile = NULL,
  outputPath = NULL,
  projectName = NULL,
  startStage = 1,
  stopStage = 2,
  numCores = 1,
  maxOrder = 2,
  numReplicates = 2,
  significanceThreshold = 0.1,
  adjustmentMethod = "fdr",
  testType = "ztest",
  orderSubset = "all"
)
}
\arguments{
\item{workspacePath}{Path to DiMSum workspace .RData file after successful completion of stage 7}

\item{inputFile}{Path to plain text file with 'WT', 'fitness', 'sigma' and 'nt_seq' or ('aa_seq' and 'STOP') columns}

\item{outputPath}{Path to directory to use for output files}

\item{projectName}{Project name}

\item{startStage}{Start at a specified pipeline stage (default:1)}

\item{stopStage}{Stop at a specified pipeline stage (default:2)}

\item{numCores}{Number of available CPU cores (default:1)}

\item{maxOrder}{Maximum number of nucleotide or amino acid substitutions for coding or non-coding sequences respectively (default:2)}

\item{numReplicates}{Number of biological replicates (or sample size) from which fitness and error estimates derived. Used to calculate degrees of freedom if 'testType' is 'ttest' (default:2)}

\item{significanceThreshold}{Significance threshold for specific and background averaged terms (default:0.1; 1 disables the treshold)}

\item{adjustmentMethod}{Method for adjusting P-values for multiple comparisons; any method in p.adjust.methods is permitted (default:'fdr')}

\item{testType}{Type of statistical test to use: either 'ztest' or 'ttest' (default:'ztest')}

\item{orderSubset}{Comma-separated list of (integer) orders of epistatic terms to retain for fitness reconstruction or 'all' (default:'all')}
}
\value{
Nothing
}
\description{
This function runs the MoCHI pipeline.
}
