% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mochi__add_fitness_to_landscapes.R
\name{mochi__add_fitness_to_landscapes}
\alias{mochi__add_fitness_to_landscapes}
\title{mochi__add_fitness_to_landscapes}
\usage{
mochi__add_fitness_to_landscapes(landscape_list, fitness_dt,
  numCores = 1)
}
\arguments{
\item{landscape_list}{list of landscape data (required)}

\item{fitness_dt}{data.table of variant fitness and error (required)}

\item{numCores}{number of available CPU cores (default:1)}
}
\value{
A list of data.tables
}
\description{
Add fitness and error to landscapes.
}
