% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/mochi__calculate_background_averaged_epistasis_terms.R
\name{mochi__calculate_background_averaged_epistasis_terms}
\alias{mochi__calculate_background_averaged_epistasis_terms}
\title{mochi__calculate_background_averaged_epistasis_terms}
\usage{
mochi__calculate_background_averaged_epistasis_terms(input_dt,
  degreesFreedom = 5, testType = "ztest",
  significanceThreshold = 0.1, adjustmentMethod = "fdr",
  numCores = 1)
}
\arguments{
\item{input_dt}{data.table as provided by calculate_individual_epistatis_terms (required)}

\item{degreesFreedom}{an integer degrees of freedom (default:5)}

\item{testType}{type of test: either "ztest" or "ttest" (default:"ztest")}

\item{significanceThreshold}{Significance threshold for specific and background averaged terms (default:0.1)}

\item{adjustmentMethod}{Method for adjusting P-values for multiple comparisons; any method in p.adjust.methods is permitted (default:'fdr')}

\item{numCores}{number of available CPU cores (default:1)}
}
\value{
A data.table with background averaged epistasis terms, direction and significance
}
\description{
Calculate general tendecies of epistasis at any order.
}
