% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mochi__calculate_individual_epistasis_terms.R
\name{mochi__calculate_individual_epistasis_terms}
\alias{mochi__calculate_individual_epistasis_terms}
\title{mochi__calculate_individual_epistasis_terms}
\usage{
mochi__calculate_individual_epistasis_terms(input_list, genotype_key,
  degreesFreedom = 5, testType = "ztest", adjustmentMethod = "fdr",
  numCores = 1)
}
\arguments{
\item{input_list}{list of combinatoral landscape data.tables with corresponding fitness and error as provided by mochi__add_fitness_to_landscapes (required)}

\item{genotype_key}{data.table with genotype codes of single substitution variants (required)}

\item{degreesFreedom}{an integer degrees of freedom (default:5)}

\item{testType}{type of test: either "ztest" or "ttest" (default:"ztest")}

\item{adjustmentMethod}{Method for adjusting P-values for multiple comparisons; any method in p.adjust.methods is permitted (default:'fdr')}

\item{numCores}{number of available CPU cores (default:1)}
}
\value{
A data.table with individual epistatic terms
}
\description{
Calculate all individual terms of epistasis.
}
