% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mochi__calculate_individual_epistatis_terms.R
\name{mochi__calculate_individual_epistatis_terms}
\alias{mochi__calculate_individual_epistatis_terms}
\title{mochi__calculate_individual_epistatis_terms}
\usage{
mochi__calculate_individual_epistatis_terms(input_list, single_mut_dt,
  degreesFreedom = 5, test_type = "ztest", numCores = 1)
}
\arguments{
\item{input_list}{list of combinatoral landscape data.tables with corresponding fitness and error as provided by mochi__add_fitness_to_landscapes (required)}

\item{single_mut_dt}{data.table of single substitution variants (required)}

\item{degreesFreedom}{an integer degrees of freedom (default:5)}

\item{test_type}{type of test: either "ztest" or "ttest" (default:"ztest")}

\item{numCores}{number of available CPU cores (default:1)}
}
\value{
A data.table with individual epistatic terms
}
\description{
Calculate all individual terms of epistasis.
}
