% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mochi__construct_var2id_list.R
\name{mochi__construct_var2id_list}
\alias{mochi__construct_var2id_list}
\title{mochi__construct_var2id_list}
\usage{
mochi__construct_var2id_list(input_dt)
}
\arguments{
\item{input_dt}{data.table of single substitution variants (required)}
}
\value{
A data.table
}
\description{
Get mapping of ids to variants for all possible combinations of single mutants supplied in data.table with id and var columns.
}
