% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mochi__create_dir.R
\name{mochi__create_dir}
\alias{mochi__create_dir}
\title{mochi__create_dir}
\usage{
mochi__create_dir(mochi_dir, execute, message = NULL,
  overwrite_dir = TRUE)
}
\arguments{
\item{mochi_dir}{directory path string (required)}

\item{execute}{whether or not the system command will be executed (required)}

\item{message}{message string (optional, default: NULL i.e. no message displayed)}

\item{overwrite_dir}{delete directory if already exists (optional, default: TRUE)}
}
\value{
Nothing
}
\description{
Create results folder for mochi pipeline stage.
}
