% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mochi__encode_genotypes.R
\name{mochi__encode_genotypes}
\alias{mochi__encode_genotypes}
\title{mochi__encode_genotypes}
\usage{
mochi__encode_genotypes(input_dt, sequenceType = "nucleotide")
}
\arguments{
\item{input_dt}{a data.table of variants (required)}

\item{sequenceType}{sequence type: one of nucleotide/aminoacid (default:nucleotide)}
}
\value{
A named list of two data.tables
}
\description{
Encode genotypes mapping id (comma-separated mutation string) to var (single character variant string where WT=0) for variable positions only.
}
