% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mochi__g_matrix.R
\name{mochi__g_matrix}
\alias{mochi__g_matrix}
\title{mochi__g_matrix}
\usage{
mochi__g_matrix(maxOrder)
}
\arguments{
\item{maxOrder}{maximum order (required)}
}
\value{
A matrix
}
\description{
Construct G matrix.
}
