% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/mochi__get_all_combinatorial_complete_landscapes.R
\name{mochi__get_all_combinatorial_complete_landscapes}
\alias{mochi__get_all_combinatorial_complete_landscapes}
\title{mochi__get_all_combinatorial_complete_landscapes}
\usage{
mochi__get_all_combinatorial_complete_landscapes(order_max = 1,
  all_variants_char, numCores = 1)
}
\arguments{
\item{order_max}{an integer number of maximum mutations to consider (default:1)}

\item{all_variants_char}{a character vector of variants each of identical length (required)}

\item{numCores}{number of available CPU cores (default:1)}
}
\value{
A list of data.tables
}
\description{
Get all combinatorial complete landscapes of all orders in all_variants_char.
}
