% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mochi__get_combinatorial_complete_landscapes.R
\name{mochi__get_combinatorial_complete_landscapes}
\alias{mochi__get_combinatorial_complete_landscapes}
\title{mochi__get_combinatorial_complete_landscapes}
\usage{
mochi__get_combinatorial_complete_landscapes(nmut = 1, all_variants_char)
}
\arguments{
\item{nmut}{an integer number of mutations to consider (default:1)}

\item{all_variants_char}{a character vector of variants each of identical character length (required)}
}
\value{
A data.table
}
\description{
Get all combinatorial complete landscapes of order nmut in all_variants_char.
}
