% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mochi__h_matrix.R
\name{mochi__h_matrix}
\alias{mochi__h_matrix}
\title{mochi__h_matrix}
\usage{
mochi__h_matrix(maxOrder)
}
\arguments{
\item{maxOrder}{maximum order (required)}
}
\value{
A matrix
}
\description{
Construct H matrix.
}
