% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mochi__reconstruct_fitness.R
\name{mochi__reconstruct_fitness}
\alias{mochi__reconstruct_fitness}
\title{mochi__reconstruct_fitness}
\usage{
mochi__reconstruct_fitness(
  input_list,
  single_mut_dt,
  order_subset,
  numCores = 1
)
}
\arguments{
\item{input_list}{list of combinatoral landscape data.tables with corresponding fitness and error as provided by mochi__add_fitness_to_landscapes (required)}

\item{single_mut_dt}{data.table of single substitution variants (required)}

\item{order_subset}{comma-separated list of (integer) orders of epistatic terms to retain for fitness reconstruction (required)}

\item{numCores}{number of available CPU cores (default:1)}
}
\value{
A data.table
}
\description{
Reconstruct fitness from individual terms of epistasis.
}
