% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mochi__string_subtract.R
\name{mochi__string_subtract}
\alias{mochi__string_subtract}
\title{mochi__string_subtract}
\usage{
mochi__string_subtract(string1, string2)
}
\arguments{
\item{string1}{string 1 (required)}

\item{string2}{string 2 (required)}
}
\value{
A string
}
\description{
Subract string2 from string1 (where strings are equal length and string2 is a zero-masked version of string1).
}
