% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mochi__ttest.R
\name{mochi__ttest}
\alias{mochi__ttest}
\title{mochi__ttest}
\usage{
mochi__ttest(av, se, df = 5, mu = 0)
}
\arguments{
\item{av}{sample mean (required)}

\item{se}{sample standard error (required)}

\item{df}{an integer degrees of freedom (default:5)}

\item{mu}{population mean (default:0)}
}
\value{
A data.table
}
\description{
TDist wrapper.
}
