% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mochi__v_matrix.R
\name{mochi__v_matrix}
\alias{mochi__v_matrix}
\title{mochi__v_matrix}
\usage{
mochi__v_matrix(maxOrder)
}
\arguments{
\item{maxOrder}{maximum order (required)}
}
\value{
A matrix
}
\description{
Construct V matrix.
}
