% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mochi__validate_input.R
\name{mochi__validate_input}
\alias{mochi__validate_input}
\title{mochi__validate_input}
\usage{
mochi__validate_input(input_list)
}
\arguments{
\item{input_list}{a list of input arguments (required)}
}
\value{
an experiment metadata object
}
\description{
Check validity and reformat input parameters.
}
