% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mochi_stage_higher_order_epistasis.R
\name{mochi_stage_higher_order_epistasis}
\alias{mochi_stage_higher_order_epistasis}
\title{mochi_stage_higher_order_epistasis}
\usage{
mochi_stage_higher_order_epistasis(dimsum_meta, epistasis_outpath,
  report = TRUE, report_outpath = NULL, save_workspace = TRUE)
}
\arguments{
\item{dimsum_meta}{an experiment metadata object (required)}

\item{epistasis_outpath}{output path for saved objects (required)}

\item{report}{whether or not to generate fitness summary plots (default: TRUE)}

\item{report_outpath}{fitness report output path}
}
\value{
Nothing
}
\description{
Obtain all epistatic terms, background relative and background averaged terms.
}
