% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mochi_stage_higher_order_epistasis_report.R
\name{mochi_stage_higher_order_epistasis_report}
\alias{mochi_stage_higher_order_epistasis_report}
\title{mochi_stage_higher_order_epistasis_report}
\usage{
mochi_stage_higher_order_epistasis_report(dimsum_meta, report_outpath)
}
\arguments{
\item{dimsum_meta}{an experiment metadata object (required)}

\item{report_outpath}{final report output path (required)}
}
\value{
Nothing
}
\description{
Generate higher order epistasis summary plots.
}
