% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mochi_stage_infer_additive_trait.R
\name{mochi_stage_infer_additive_trait}
\alias{mochi_stage_infer_additive_trait}
\title{mochi_stage_infer_additive_trait}
\usage{
mochi_stage_infer_additive_trait(mochi_meta, additivetrait_outpath,
  report = TRUE, report_outpath = NULL, save_workspace = TRUE)
}
\arguments{
\item{mochi_meta}{an experiment metadata object (required)}

\item{additivetrait_outpath}{output path for saved objects (required)}

\item{report}{whether or not to generate fitness summary plots (default: TRUE)}

\item{report_outpath}{fitness report output path}

\item{save_workspace}{whether or not to save the current workspace (default: TRUE)}
}
\value{
Nothing
}
\description{
Infer additive trait using artifical neural network.
}
