% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mochi_stage_reconstruct_fitness.R
\name{mochi_stage_reconstruct_fitness}
\alias{mochi_stage_reconstruct_fitness}
\title{mochi_stage_reconstruct_fitness}
\usage{
mochi_stage_reconstruct_fitness(mochi_meta, reconstruct_outpath,
  report = TRUE, report_outpath = NULL, save_workspace = TRUE)
}
\arguments{
\item{mochi_meta}{an experiment metadata object (required)}

\item{reconstruct_outpath}{output path for saved objects (required)}

\item{report}{whether or not to generate fitness summary plots (default: TRUE)}

\item{report_outpath}{fitness report output path}

\item{save_workspace}{whether or not to save the current workspace (default: TRUE)}
}
\value{
Nothing
}
\description{
Reconstruct fitness from individual terms of epistasis.
}
