% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mochi.R
\name{mochi}
\alias{mochi}
\title{Run MoCHI pipeline}
\usage{
mochi(workspacePath, inputFile, outputPath, projectName, startStage = 1,
  stopStage = 0, numCores = 1, maxOrder = 2, numReplicates = 2,
  FDR_threshold = 0.1)
}
\arguments{
\item{workspacePath}{Path to DiMSum workspace .RData file after successful completion of stage 7}

\item{inputFile}{Path to plain text file with 'WT', 'fitness', 'sigma' and 'nt_seq' or ('aa_seq' and 'STOP') columns}

\item{outputPath}{Path to directory to use for output files}

\item{projectName}{Project name}

\item{startStage}{Start at a specified pipeline stage (default:1)}

\item{stopStage}{Stop at a specified pipeline stage (default:0 i.e. no stop condition)}

\item{numCores}{Number of available CPU cores (default:1)}

\item{maxOrder}{Maximum number of nucleotide or amino acid substitutions for coding or non-coding sequences respectively (default:2)}

\item{numReplicates}{Number of biological replicates (or sample size) from which fitness and error estimates derived (default:2)}

\item{FDR_threshold}{FDR threshold for significant specific and background averaged terms (default:0.1)}
}
\value{
Nothing
}
\description{
This function runs the MoCHI pipeline.
}
