% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mochi.R
\name{mochi}
\alias{mochi}
\title{Run MoCHI pipeline}
\usage{
mochi(outputPath, projectName, startStage = 1, stopStage = 0,
  numCores = 1, maxOrder = 2)
}
\arguments{
\item{outputPath}{Path to directory to use for output files}

\item{projectName}{Project name}

\item{startStage}{Start at a specified pipeline stage (default:1)}

\item{stopStage}{Stop at a specified pipeline stage (default:0 i.e. no stop condition)}

\item{numCores}{Number of available CPU cores (default:1)}

\item{maxOrder}{Maximum number of nucleotide or amino acid substitutions for coding or non-coding sequences respectively (default:2)}

\item{workspacePath}{Path to DiMSum workspace .RData file after successful completion of stage 7}

\item{inputFile}{Path to plain text file with 'WT', 'fitness', 'sigma' and 'nt_seq' or ('aa_seq' and 'STOP') columns}
}
\value{
Nothing
}
\description{
This function runs the MoCHI pipeline.
}
